split out rsl functions; only build rsl2odim when rsl detected
