+++ configure.in
@@ -430,7 +430,7 @@ AC_SUBST(RSL_INCLUDE_FLAG)
 AC_SUBST(RSL_LIBRARY_FLAG)
 AC_SUBST(RSL_CFLAG)
 AC_SUBST(RSL_LIB)
-AC_SUBST(LD_PRINTOUT)
+AC_SUBST(GOT_RSL_LIB)
 AC_SUBST(LD_PRINTOUT)
 
 AC_CONFIG_FILES(def.mk)
