+++ lib/Makefile
@@ -26,7 +26,7 @@ SRC_VOL2BIRD_DIR = .
 
 all : libvol2bird.so
 
-LIBVOL2BIRD_DEPS = constants.h libsvdfit.h libdealias.h libvol2bird.h libsvdfit.c libdealias.c libvol2bird.c
+LIBVOL2BIRD_DEPS = constants.h libsvdfit.h libdealias.h librsl.h libvol2bird.h libsvdfit.c libdealias.c librsl.c libvol2bird.c
 
 libvol2bird.so : $(LIBVOL2BIRD_DEPS)
 	# ------------------------------------
@@ -46,6 +46,7 @@ libvol2bird.so : $(LIBVOL2BIRD_DEPS)
 	$(SRC_VOL2BIRD_DIR)/libvol2bird.c \
 	$(SRC_VOL2BIRD_DIR)/libsvdfit.c \
 	$(SRC_VOL2BIRD_DIR)/libdealias.c \
+	$(SRC_VOL2BIRD_DIR)/librsl.c \
 	$(LDFLAGS) \
 	-Wall -o libvol2bird.so $(RAVE_MODULE_LIBRARIES) -lconfuse -lgsl -lgslcblas $(RSL_LIB) -lm
