+++ lib/constants.h
@@ -96,9 +96,9 @@
 // Name of the program, to be stored as task attribute in ODIM
 #define PROGRAM "vol2bird"
 // Version of the program, to be stored as task_version attribute in ODIM
-#define VERSION "0.3.13"
+#define VERSION "0.3.14"
 // Date of latest version of the program
-#define VERSIONDATE "4-May-2017"
+#define VERSIONDATE "5-May-2017"
 
 //-------------------------------------------------------//
 //  user options defaults (to be set in options.conf)    //
