+++ lib/librsl.h
@@ -1,0 +1,6 @@
+#ifdef RSL
+
+
+PolarVolume_t* vol2birdGetRSLVolume(char* filename, float rangeMax, int small);
+
+#endif
